extracted weather request to hook
